Configure the logger for console friendliness
The default env_logger is nice for logs, but it feels a bit out of place
for interactive console apps.  Initialize it with a more "traditional"
style and make sure the default is to show INFO-level messages too.

While doing this, allow the app to disable automatic initialization of
the logger because some apps (e.g. those that do daemonization) may want
to postpone initialization.